Add LibFuzzer support
Nginx can be built as a fuzzer target using the `--with-fuzzer`
configure flag. This feature is currently only supported in Clang, thus
you may need to explicitly set the compiler (e.g. `--with-cc=clang`).

The fuzzer will use the `nginx-fuzzer.conf` installed into the conf
directory instead of the default `nginx.conf`. This config file must
listen on the unix domain socket `nginx.sock` in the current working
directory.